pyproject.toml
[project]
name = "comic-exp"
version = "0.1.0"
description = "Pipeline for generating storyboards from narrative text"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "transformers",
    "torch",
    "diffusers",
    "accelerate",
    "sentencepiece",
    "spacy",
    "pillow",
    "gradio",
    "langdetect",
    "tiktoken",
    "protobuf",
    "pytest>=9.0.2",
]

[project.scripts]
"comic-exp" = "comic_exp:main"
